requirements.txt
streamlit
python-dotenv
euriai
pandas
langchain
fastmcp
fitz
youtube-transcript-api
PyMuPDF